Create deck form; update deck CRUD routes

+++ api/decks.py
@@ -1,12 +1,15 @@
 import json
 
-from django.shortcuts import render
+from django.shortcuts import get_object_or_404
 from django.http import JsonResponse
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_http_methods
+from django.core.exceptions import PermissionDenied
 from django.contrib.auth.models import User
+from django.forms.models import model_to_dict
 
 from .models import Deck
+from .forms import DeckForm
 
 # Create your views here.
 ## controllers for DECKS
@@ -20,26 +23,43 @@ def list_decks(request):
 
 
 @require_http_methods(['POST'])
-def create_deck(request, user_pk):
-    new_deck_info = json.loads(request.body)
-    deck = Deck.objects.create(
-        **new_deck_info,
-        user=User.objects.get(pk=user_pk)
-    )
-    return JsonResponse({'deck': new_deck_info}, status=201)
+@login_required
+def create_deck(request):
+    form = DeckForm(json.loads(request.body))
+    if form.is_valid():
+        deck = form.save(commit=False)
+        deck.user = request.user
+        deck.save()
+        return JsonResponse({'deck': model_to_dict(deck)}, status=201)
+    else:
+        return JsonResponse({'error': 'invalid form input'}, status=400)
 
 
 @require_http_methods(['PUT'])
-def edit_deck(request, user_pk, deck_pk):
-    deck = Deck.objects.get(pk=deck_pk)
-    new_info = json.loads(request.body)
-
-    Deck.objects.filter(pk=deck_pk).update(**new_info)
-    return JsonResponse(new_info, status=200)
+@login_required
+def edit_deck(request, deck_pk):
+    deck = get_object_or_404(
+        Deck.objects.select_related('user'),
+        pk=deck_pk
+    )
+    if deck.user != request.user:
+        raise PermissionDenied
+    form = DeckForm(json.loads(request.body), instance=deck)
+    if form.is_valid():
+        deck = form.save()
+        return JsonResponse({'deck': model_to_dict(deck)}, status=200)
+    else:
+        return JsonResponse({'error': 'edit error'}, status=400)
 
 
 @require_http_methods(['DELETE'])
-def delete_deck(request, user_pk, deck_pk):
-    deck = Deck.objects.get(pk=deck_pk)
+@login_required
+def delete_deck(request, deck_pk):
+    deck = get_object_or_404(
+        Deck.objects.select_related('user'),
+        pk=deck_pk
+    )
+    if deck.user != request.user:
+        raise PermissionDenied
     deck.delete()
-    return JsonResponse({}, status=200)
+    return JsonResponse({'message': 'delete successful'}, status=200)
